lambdas/publish_to_linkedin/requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
oauthlib==2.0.6
python3-linkedin==1.0.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
